moved queries and changed names
