+++ src/utils/Query/CompanyQuery.js
@@ -44,25 +44,6 @@ const GET_COMPANY = gql`
   }
 `
 
-const GET_USER_PREF = gql`
-  query getPref($user_id: uuid!) {
-    Preferences(where: { user_id: { _eq: $user_id } }) {
-      locals
-      current_company
-    }
-  }
-`
-
-const SET_LOCALS = gql`
-  mutation setLocals($user_id: uuid!, $locals: String!) {
-    update_Preferences(
-      where: { user_id: { _eq: $user_id } }
-      _set: { locals: $locals }
-    ) {
-      affected_rows
-    }
-  }
-`
 
 const SET_ACTIVE_COMPANY = gql`
   mutation setCompany($user_id: uuid!, $current_company: String!) {
@@ -128,27 +109,7 @@ const GET_SUBSCRIP_COMPANY = gql`
     }
   }
 `
-
-const GET_SUBSCRIP_ACCOUNTS = gql`
-  subscription($company_id: uuid!) {
-    Account(where: { company_id: { _eq: $company_id } }) {
-      id
-      name
-      debit
-      balance
-      company_id
-      account_numbers
-      Company {
-        user_id
-        name
-      }
-    }
-  }
-`
-
-
-
-const CREATE_COMPANY = gql`
+const POST_COMPANY = gql`
   mutation createCompany($name: String!, $mother_id: uuid, $user_id: uuid!) {
     insert_Company(
       objects: { name: $name, mother_id: $mother_id, user_id: $user_id }
@@ -158,44 +119,5 @@ const CREATE_COMPANY = gql`
   }
 `
 
-const CREATE_ACCOUNT = gql`
-  mutation createAccount(
-    $name: String!
-    $balance: numeric!
-    $debit: Boolean!
-    $company_id: uuid!
-  ) {
-    insert_Account(
-      objects: {
-        name: $name
-        balance: $balance
-        debit: $debit
-        company_id: $company_id
-      }
-    ) {
-      affected_rows
-    }
-  }
-`
-const GET_DAY_BOOK = gql`
-  subscription($company_id: uuid!) {
-    day_book(where: { company_id: { _eq: $company_id } }) {
-      account_id
-      name
-      debit
-      balance
-    }
-  }
-`
 
-export {
-  GET_COMPANY,
-  GET_DAY_BOOK,
-  GET_SUBSCRIP_COMPANY,
-  GET_SUBSCRIP_ACCOUNTS,
-  CREATE_ACCOUNT,
-  CREATE_COMPANY,
-  GET_USER_PREF,
-  SET_LOCALS,
-  SET_ACTIVE_COMPANY,
-}
+export {GET_COMPANY, SET_ACTIVE_COMPANY, GET_SUBSCRIP_COMPANY, POST_COMPANY}
